% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark}
\alias{benchmark}
\title{Execute the benchmark}
\usage{
benchmark(
  forecaster,
  usecase,
  type = "one",
  output = "benchmark.csv",
  name = "Benchmarked Method"
)
}
\arguments{
\item{forecaster}{The forecasting method. This method gets a timeseries objekt (ts) and the horizon (h). The method returns the forecast values.}

\item{usecase}{The use case for the benchmark. It must be either economics, finance, human, or nature.}

\item{type}{Optional parameter: The evaluation type. It must be either one (one-step-ahead forecast), multi (multi-step-ahead forecast), or rolling (rolling-origin forecast). one by default.}

\item{output}{Optional parameter: The name of the output file. benchmark.csv by default.}

\item{name}{Optional parameter: The name of the forecasting method. Benchmarked Method by default.}
}
\value{
The performance of the forecasting method in comparison with state-of-the-art methods.
}
\description{
Automatically evaluates and ranks forecasting methods based on their performance.
}
\examples{
# Example usage
benchmark(forecaster,usecase="economics",type="one")

# Example forecasting method
forecaster <- function(ts,h){ return(forecast(ets(ts), h = h)$mean) }
}
\author{
Andre Bauer
}
