% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{calculateMeasures}
\alias{calculateMeasures}
\title{Calculating the performance measures}
\usage{
calculateMeasures(end, start, hist, forecast, actual, horizon)
}
\arguments{
\item{end}{The timestamp after the forecast.}

\item{start}{The timestamp before the forecast.}

\item{hist}{The training data.}

\item{forecast}{The forecast.}

\item{actual}{The test data.}

\item{horizon}{The forecast horizon.}
}
\value{
The performance measures of the forecast.
}
\description{
Calculates all performance measures.
}
