% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{evaluation}
\alias{evaluation}
\title{Performing the forecasts}
\usage{
evaluation(forecaster, data, type)
}
\arguments{
\item{forecaster}{The forecasting method. This method gets a timeseries objekt (ts) and the horizon (h). The method returns the forecast values.}

\item{type}{The evaluation type. It must be either one (one-step-ahead forecast), multi (multi-step-ahead forecast), or rolling (rolling-origin forecast). one by default.}

\item{usecase}{The use case for the benchmark. It must be either economics, finance, human, or nature.}
}
\value{
The performance measures of the forecasting method.
}
\description{
Peforms the forecast of the forecasting method.
}
