% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{forecast.one}
\alias{forecast.one}
\title{Performing the one-step-ahead forecast}
\usage{
\method{forecast}{one}(forecaster, timeseries)
}
\arguments{
\item{forecaster}{The forecasting method. This method gets a timeseries objekt (ts) and the horizon (h). The method returns the forecast values.}

\item{timeseries}{The time series.}
}
\value{
The performance measures of the one-step-ahead forecast of the time series.
}
\description{
Peforms the one-step-ahead forecast of a time series based on the forecasting method.
}
