% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{forecast.rolling}
\alias{forecast.rolling}
\title{Performing the one-step-ahead forecast}
\usage{
\method{forecast}{rolling}(forecaster, timeseries)
}
\arguments{
\item{forecaster}{The forecasting method. This method gets a timeseries objekt (ts) and the horizon (h). The method returns the forecast values.}

\item{timeseries}{The time series.}
}
\value{
The performance measures of the rolling-origin forecast of the time series.
}
\description{
Peforms the rolling-origin forecast of a time series based on the forecasting method.
}
