% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{getindizies}
\alias{getindizies}
\title{Determining the indices.}
\usage{
getindizies(timeseries)
}
\arguments{
\item{timeseries}{The time series.}
}
\value{
The indices.
}
\description{
Determines the indices for the rolling-origin forecast.
}
