% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{mase}
\alias{mase}
\title{Calculating the MASE of the forecast}
\usage{
mase(forecast, actual, hist)
}
\arguments{
\item{forecast}{The forecast values}

\item{actual}{The test data.}

\item{hist}{The training data.}
}
\value{
The MASE of the forecast
}
\description{
Calculates the Mean Absolute Scaled Error (MASE) between the forecast and test data.
}
