% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{moas}
\alias{moas}
\title{Calculating the MOAS of the forecast}
\usage{
moas(forecast, actual)
}
\arguments{
\item{forecast}{The forecast values}

\item{actual}{The test data.}
}
\value{
The MOAS of the forecast
}
\description{
Calculates the Mean Over-Accuracy Share (MOAS) between the forecast and test data.
}
