% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{moes}
\alias{moes}
\title{Calculating the MOES of the forecast}
\usage{
moes(forecast, actual)
}
\arguments{
\item{forecast}{The forecast values}

\item{actual}{The test data.}
}
\value{
The MOES of the forecast
}
\description{
Calculates the Mean Over-Estimation Share (MOES) between the forecast and test data.
}
