% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{muas}
\alias{muas}
\title{Calculating the MUAS of the forecast}
\usage{
muas(forecast, actual)
}
\arguments{
\item{forecast}{The forecast values}

\item{actual}{The test data.}
}
\value{
The MUAS of the forecast
}
\description{
Calculates the Mean Under-Accuracy Share (MUAS) between the forecast and test data.
}
