% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{mues}
\alias{mues}
\title{Calculating the MUES of the forecast}
\usage{
mues(forecast, actual)
}
\arguments{
\item{forecast}{The forecast values}

\item{actual}{The test data.}
}
\value{
The MUES of the forecast
}
\description{
Calculates the Mean Under-Estimation Share (MUES) between the forecast and test data.
}
