% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{smape}
\alias{smape}
\title{Calculating the sMAPE of the forecast}
\usage{
smape(forecast, actual)
}
\arguments{
\item{forecast}{The forecast values}

\item{actual}{The test data.}
}
\value{
The sMAPE of the forecast
}
\description{
Calculates the Symmetrical Mean Absolute Percentage Error (sMAPE) between the forecast and test data.
}
